更新 'hexo/hexo.sh'

+++ hexo/hexo.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
-workDir=/blog
-cd $workDir
-hexo init 
-mkdir /blog/config
-mv /blog/_config.yml /blog/config/config.yml
-  # install plugins for hexo
+if [ ! -f "/blog/config/config.yml" ];then
+	mkdir -p /temp
+	cd /temp
+	hexo init
+	cp -r /temp/* /blog/ 
+	mv /blog/_config.yml /blog/config/config.yml
+	cd /blog
+	rm -rf /temp
+fi
 hexo g -dw --config /blog/config/config.yml
+
